TGAM1/requirements.txt
pyserial
matplotlib
PyOpenGL
glfw